Eu adicionei a verificação de se ja tava logado e esqueci que todos os testes ignoram esse fato, portanto adicionei o logout para conseguir sair da sessão atual.

+++ src/test/java/FilePersistentTest.java
@@ -3,6 +3,7 @@ import balbucio.org.firebase4j.FirebaseOptions;
 import balbucio.org.firebase4j.model.User;
 import balbucio.org.firebase4j.persistent.FilePersistent;
 import org.junit.jupiter.api.*;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.File;
@@ -16,7 +17,7 @@ public class FilePersistentTest {
     private User createdUser;
 
     @BeforeAll
-    public void create() throws Exception{
+    public void create() throws Exception {
         options = FirebaseOptions.fromJsonFile(new File("test-credentials.json"));
         options.setPersistent(new FilePersistent(new File("firebase.persistent")));
         auth = FirebaseAuth.newInstance(options);
@@ -25,29 +26,31 @@ public class FilePersistentTest {
     @Test
     @DisplayName("Criar e salvar conta")
     @Order(0)
-    public void createAndSave() throws Exception{
+    public void createAndSave() throws Exception {
         createdUser = auth.signInAnonymously();
     }
 
     @Test
     @DisplayName("Checar se a conta foi salva")
     @Order(1)
-    public void check() throws Exception{
+    public void check() throws Exception {
         assertNotNull(auth.getCurrentUser());
     }
 
     @Test
     @DisplayName("Checar em uma instância diferente")
     @Order(2)
-    public void recreate() throws Exception{
+    public void recreate() throws Exception {
         FirebaseAuth auth2 = FirebaseAuth.newInstance(options);
         assertNotNull(auth2.getCurrentUser());
-        assertTrue(auth2.getCurrentUser().getIdToken().equalsIgnoreCase(createdUser.getIdToken()));
+        assertTrue(auth2.getCurrentUser().getLocalId().equalsIgnoreCase(createdUser.getLocalId()));
     }
 
     @BeforeAll
-    public void deleteUser() throws Exception{
-        createdUser.delete();
+    public void deleteUser() throws Exception {
+        if (createdUser != null) {
+            createdUser.delete();
+        }
         options.getPersistent().clear();
     }
